pd_frames/requirements.txt
fastapi
fastapi-sqlalchemy
SQLAlchemy
pyodbc
uvicorn[standard]
python-decouple
behave
yml
opencv-python
cython
wget
decord
minio
Pillow
python-multipart
validators
aiohttp
aiofiles
pathlib